Move enrollment admission into enrollment library
