+++ apps/prairielearn/src/lib/enrollment/eligibility.ts
@@ -1,6 +1,6 @@
 import { assertNever } from '@prairielearn/utils';
 
-import type { Course, CourseInstance, Enrollment, User } from './db-types.js';
+import type { Course, CourseInstance, Enrollment, User } from '../db-types.js';
 
 export type EnrollmentIneligibilityReason =
   | 'blocked'
